pyproject.toml
[project]
name = "telegram-filter-notifai"
version = "0.1.0"
description = "A smart tool for filtering and notifying relevant Telegram messages based on custom criteria, leveraging AI to streamline your Telegram experience"
authors = [
    { name = "Leynier Gutiérrez González", email = "leynier41@gmail.com" },
]
readme = "README.md"

requires-python = ">=3.11"

dependencies = [
    "pydbantic[sqlite]>=0.0.47",
    "openai>=1.42.0",
    "pyrogram>=2.0.106",
    "tgcrypto>=1.2.5",
    "python-dotenv>=1.0.1",
    "uvloop>=0.20.0",
    "typer>=0.12.4",
]

[tool.uv]
dev-dependencies = ["ruff>=0.6.2"]

[project.scripts]
telegram-filter-notifai = "telegram_filter_notifai.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
